requirements.txt
chromadb
langchain_community
langchain_experimental
langchain_huggingface
langchain_groq
langchain_openai
langchain_chroma
langchain_core
langsmith
tqdm
python-dotenv
pandas
termcolor
pypdf
unstructured
python-docx
flask
flask-cors
gunicorn
langgraph